refactor(shared-log): extract the coordinate persistence module (stage-4.5)
The 39 coordinate-persistence methods (34 main-cluster + 5 early helpers)
move from SharedLog onto the CoordinatePersistenceCoordinator with
byte-identical bodies, verified by diffing the moved text against the base
(5c4568e32) bodies under the documented glue mapping:
`this._coordinates.<field>` -> `this.<field>` for the four owned fields and
`this.<hostDep>` -> `this.deps.<hostDep>()` for host state, plus exactly
three mechanical extras — the `_nativeDurableRecoveryReadyForReopen = true`
write becomes `deps.setDurableRecoveryReadyForReopen(true)`, the three
`new this.indexableDomain.constructorEntry(...)` sites gain parentheses
around the member expression (`new` binds call parens otherwise), and the
two persistCoordinatesBatch native-commit blocks capture the deps value in
a block-local (the blocks are synchronous, so one read at entry is
observationally identical to the legacy per-access field reads, and TS
cannot narrow across calls).

Every dep is a late-bound closure into the live host (native runtimes,
coordinate index, remoteBlocks, coordinateToHash, timeUntilRoleMaturity,
durability/drop flags are all re-assigned across open/close/drop cycles);
the ownership-lifecycle helpers stay host methods. The prepared-join commit
closures returned by createNativeBackbonePreparedJoinCommit still route
their lifecycle/poison checks to host state through deps, so the poison
surface is unchanged. ensureCurrentHeadCoordinatesIndexed stays host-side
(it needs the leader planner); _nativeBackboneDropStarted stays host with a
deps closure.

SharedLog keeps thin KEEP-OLD delegators for all 29 externally-called names
(the plan's 16 plus the 13 additional names the call-site census proved are
called from host code: createCoordinatePersistenceEntryFromNativePlan,
materializePreparedCoordinateEntry, materializeResidentCoordinateEntry,
materializeRepairDispatchEntries, snapshotResidentCoordinateEntries,
rollbackNativeBackboneCoordinateAppend{,Durably},
persistPreparedCoordinateNativeTransaction,
persistBackboneCoordinateFieldsNativeTransaction,
canUseBackboneOnlyCoordinatePersistence,
canUseNativeBackboneResidentCoordinateState,
forget{,Resident}CoordinateStateForHashes), so every legacy call site is
untouched; the 10 internal-only helpers moved without delegators. The
maybe-promise idiom and the delete-hash helpers moved with the methods
(index.ts re-imports them); the coordinate type declarations stay in
index.ts and are exported for the module's type-only imports (no runtime
cycle: the compiled coordinate-persistence.js does not import index.js).

The host `_coordinates` field is typed `<any>` deliberately: the deps
closures mention R in invariant positions, and an R-typed field makes
SharedLog instantiations invariant in R; the compat accessors and
delegators re-state the precise R types at every boundary. That same
type-graph reshaping made the compiler actually relate the two huge
EventStore instantiations in replicate.spec.ts (it previously gave up at
depth and let a mismatched annotation through) — the declarations now match
the instantiations. Test spies/stubs of moved method names re-point to the
coordinator (`_coordinates`), which every call — delegated or
coordinator-internal — passes through, preserving legacy interception
exactly (the receive-admission-pins precedent for re-pointing internal
probes to a fence's new home).
