+++ packages/programs/data/shared-log/src/index.ts
@@ -61,11 +61,8 @@ import {
 } from "@peerbit/log";
 import { logger as loggerFn } from "@peerbit/logger";
 import type {
-	NativeBackboneAppendProfile,
 	NativeBackboneAppendResult,
-	NativeBackboneCoordinateCommitColumns,
 	NativeBackboneCoordinateFields,
-	NativeBackboneLogCommitEntry,
 	NativeBackboneRawReceiveGroupAssignmentPlan,
 	NativeBackboneRawReceiveGroupIndexPlan,
 	NativeBackboneRawReceiveGroupLeaderPlan,
@@ -181,7 +178,14 @@ import {
 	materializeVerifiedRawExchangeHeadsMessage,
 } from "./exchange-heads.js";
 import { FanoutEnvelope } from "./fanout-envelope.js";
-import { CoordinatePersistenceCoordinator } from "./coordinate-persistence.js";
+import {
+	CoordinatePersistenceCoordinator,
+	type MaybePromise,
+	combineCoordinateDeleteHashes,
+	isPromiseLike,
+	mapMaybePromise,
+	normalizedHashValues,
+} from "./coordinate-persistence.js";
 import { InstanceLifecycle } from "./instance-lifecycle.js";
 import {
 	MAX_U32,
@@ -290,16 +294,6 @@ type SharedLogServicesWithFanout = {
 
 const getSharedLogFanoutService = (services: unknown): FanoutTree | undefined =>
 	(services as SharedLogServicesWithFanout).fanout;
-
-type MaybePromise<T> = T | Promise<T>;
-
-const isPromiseLike = <T>(value: MaybePromise<T>): value is Promise<T> =>
-	!!value && typeof (value as Promise<T>).then === "function";
-
-const mapMaybePromise = <T, R>(
-	value: MaybePromise<T>,
-	fn: (value: T) => MaybePromise<R>,
-): MaybePromise<R> => (isPromiseLike(value) ? value.then(fn) : fn(value));
 
 type PendingIHave<T> = {
 	resetTimeout: () => void;
@@ -520,7 +514,7 @@ type ReplicationRangeDeletionOutcome<R extends "u32" | "u64"> = {
 	error?: unknown;
 };
 
-type EntryLeaderPlan<R extends "u32" | "u64"> = {
+export type EntryLeaderPlan<R extends "u32" | "u64"> = {
 	coordinates: NumberFromType<R>[];
 	coordinateStrings?: string[];
 	leaders: LeaderMap;
@@ -528,13 +522,13 @@ type EntryLeaderPlan<R extends "u32" | "u64"> = {
 	assignedToRangeBoundary?: boolean;
 };
 
-type ReusableReceiveCoordinatePlan<R extends "u32" | "u64"> = {
+export type ReusableReceiveCoordinatePlan<R extends "u32" | "u64"> = {
 	plan: EntryLeaderPlan<R>;
 	replicas: number;
 	prepared: PreparedCoordinatePersistence<R>;
 };
 
-type DecodedReplicaCountMap = ReadonlyMap<string, number>;
+export type DecodedReplicaCountMap = ReadonlyMap<string, number>;
 export type SharedLogCoordinateNativeFields<R extends "u32" | "u64"> = {
 	hash: string;
 	hashNumber: NumberFromType<R>;
@@ -548,7 +542,7 @@ export type SharedLogCoordinateNativeFields<R extends "u32" | "u64"> = {
 	metaBytes: Uint8Array;
 };
 
-type PreparedCoordinatePersistence<R extends "u32" | "u64"> = {
+export type PreparedCoordinatePersistence<R extends "u32" | "u64"> = {
 	coordinateEntry?: EntryReplicated<R>;
 	assignedToRangeBoundary: boolean;
 	fields: SharedLogCoordinateNativeFields<R>;
@@ -558,7 +552,7 @@ export type ResidentCoordinateEntry<R extends "u32" | "u64"> =
 	| EntryReplicated<R>
 	| SharedLogCoordinateNativeFields<R>;
 
-type CoordinatePersistBatchItem<R extends "u32" | "u64"> = {
+export type CoordinatePersistBatchItem<R extends "u32" | "u64"> = {
 	coordinates: NumberFromType<R>[];
 	entry: ShallowOrFullEntry<any> | EntryReplicated<R>;
 	leaders: LeaderMap | false;
@@ -571,25 +565,25 @@ type CoordinatePersistBatchItem<R extends "u32" | "u64"> = {
 	prepared?: PreparedCoordinatePersistence<R>;
 };
 
-type NativeBackboneReceiveCoordinateRow<R extends "u32" | "u64"> = {
+export type NativeBackboneReceiveCoordinateRow<R extends "u32" | "u64"> = {
 	item: CoordinatePersistBatchItem<R>;
 	prepared: PreparedCoordinatePersistence<R>;
 	fields: SharedLogCoordinateNativeFields<R>;
 	deleteHashes: string[];
 };
 
-type NativeBackboneReceiveCoordinateBatch<R extends "u32" | "u64"> = {
+export type NativeBackboneReceiveCoordinateBatch<R extends "u32" | "u64"> = {
 	rows: NativeBackboneReceiveCoordinateRow<R>[];
 	rollbackCoordinateEntries?: NativeBackboneCoordinateRollback<R>;
 };
 
-type NativeBackboneCoordinateRollback<R extends "u32" | "u64"> = {
+export type NativeBackboneCoordinateRollback<R extends "u32" | "u64"> = {
 	hashes: Set<string>;
 	entries: Map<string, ResidentCoordinateEntry<R>>;
 	generations: Map<string, number>;
 };
 
-type RepairDispatchEntry<R extends "u32" | "u64"> = ResidentCoordinateEntry<R>;
+export type RepairDispatchEntry<R extends "u32" | "u64"> = ResidentCoordinateEntry<R>;
 
 type PreparedLocalAppendCommit<R extends "u32" | "u64"> = {
 	hash: string;
@@ -1049,7 +1043,7 @@ const isReplicationOptionsDependentOnPreviousState = async (
 };
 
 
-interface IndexableDomain<R extends "u32" | "u64"> {
+export interface IndexableDomain<R extends "u32" | "u64"> {
 	numbers: Numbers<R>;
 	constructorEntry: new (properties: {
 		coordinates: NumberFromType<R>[];
@@ -1072,7 +1066,7 @@ interface IndexableDomain<R extends "u32" | "u64"> {
 	) => ReplicationRangeIndexable<R>;
 }
 
-type PutAndDeleteIndex<T extends Record<string, any>> = Index<T> & {
+export type PutAndDeleteIndex<T extends Record<string, any>> = Index<T> & {
 	putAndDelete?: (
 		value: T,
 		deleteOptions: DeleteOptions,
@@ -1143,59 +1137,9 @@ type PutAndDeleteIndex<T extends Record<string, any>> = Index<T> & {
 	) => Promise<unknown> | unknown;
 };
 
-type EntryWithMetaBytes = {
+export type EntryWithMetaBytes = {
 	getMetaBytes?: () => Uint8Array | undefined;
 	getHashDigestBytes?: () => Uint8Array | undefined;
-};
-
-const EMPTY_HASHES: string[] = [];
-
-const normalizedHashValues = (hashes: Iterable<string>): string[] => {
-	if (Array.isArray(hashes)) {
-		if (hashes.length === 0) {
-			return EMPTY_HASHES;
-		}
-		if (hashes.length === 1) {
-			return hashes[0] ? hashes : EMPTY_HASHES;
-		}
-	}
-	const values: string[] = [];
-	const seen = new Set<string>();
-	for (const hash of hashes) {
-		if (!hash || seen.has(hash)) {
-			continue;
-		}
-		seen.add(hash);
-		values.push(hash);
-	}
-	return values;
-};
-
-const combineCoordinateDeleteHashes = (
-	nextHashes: string[],
-	deleteHashes?: string[],
-): string[] => {
-	if (!deleteHashes || deleteHashes.length === 0) {
-		return nextHashes;
-	}
-	if (nextHashes.length === 0) {
-		return deleteHashes;
-	}
-	const combined: string[] = [];
-	const seen = new Set<string>();
-	for (const hash of nextHashes) {
-		if (!seen.has(hash)) {
-			seen.add(hash);
-			combined.push(hash);
-		}
-	}
-	for (const hash of deleteHashes) {
-		if (!seen.has(hash)) {
-			seen.add(hash);
-			combined.push(hash);
-		}
-	}
-	return combined;
 };
 
 const createIndexableDomainFromResolution = <R extends "u32" | "u64">(
@@ -1957,7 +1901,13 @@ export class SharedLog<
 	// site where the legacy `??=` defaults ran); all internal readers use
 	// direct property hops (`this._coordinates.<field>`), never the compat
 	// accessors below.
-	private _coordinates!: CoordinatePersistenceCoordinator<R>;
+	// Typed `<any>` deliberately: the coordinator's deps closures mention R in
+	// invariant positions, and an R-typed field here would make SharedLog
+	// instantiations invariant in R (breaking the long-standing
+	// `EventStore<string, any>` assignability tests rely on). The compat
+	// accessors and KEEP-OLD delegators below re-state the precise R types at
+	// every boundary, so host-side inference is unchanged.
+	private _coordinates!: CoordinatePersistenceCoordinator<any>;
 	// Test-visible compat accessors under the historical field names. The
 	// getters mirror the legacy never-opened-clone reads (`undefined` before
 	// the coordinator exists); the setters hydrate the coordinator exactly
@@ -1972,7 +1922,7 @@ export class SharedLog<
 		value: Map<string, ResidentCoordinateEntry<R>> | undefined,
 	) {
 		(this._coordinates ??=
-			new CoordinatePersistenceCoordinator<R>())._residentEntryCoordinatesByHash =
+			this.createCoordinatePersistenceCoordinator())._residentEntryCoordinatesByHash =
 			value;
 	}
 
@@ -1986,7 +1936,7 @@ export class SharedLog<
 		value: NativeBackboneCoordinatePersistenceAdapter | undefined,
 	) {
 		(this._coordinates ??=
-			new CoordinatePersistenceCoordinator<R>())._nativeBackboneCoordinatePersistence =
+			this.createCoordinatePersistenceCoordinator())._nativeBackboneCoordinatePersistence =
 			value;
 	}
 
@@ -1998,7 +1948,7 @@ export class SharedLog<
 		value: Map<string, number> | undefined,
 	) {
 		(this._coordinates ??=
-			new CoordinatePersistenceCoordinator<R>())._nativeCoordinateMutationGenerations =
+			this.createCoordinatePersistenceCoordinator())._nativeCoordinateMutationGenerations =
 			value;
 	}
 
@@ -2008,7 +1958,7 @@ export class SharedLog<
 
 	set _nativeBackboneCoordinateJournalLastFlushMs(value: number) {
 		(this._coordinates ??=
-			new CoordinatePersistenceCoordinator<R>())._nativeBackboneCoordinateJournalLastFlushMs =
+			this.createCoordinatePersistenceCoordinator())._nativeBackboneCoordinateJournalLastFlushMs =
 			value;
 	}
 
@@ -3585,6 +3535,49 @@ export class SharedLog<
 		this._instanceLifecycle = this.createInstanceLifecycle();
 	}
 
+	// Every dep is a late-bound closure into the live host: the native
+	// runtimes, the coordinate index, remoteBlocks, coordinateToHash,
+	// timeUntilRoleMaturity, and the durability/drop flags are all
+	// re-assigned across open/close/drop cycles, so nothing is captured by
+	// value; the ownership-lifecycle helpers stay host methods (sinon spies
+	// and the InstanceLifecycle fold depend on them).
+	private createCoordinatePersistenceCoordinator(): CoordinatePersistenceCoordinator<R> {
+		return new CoordinatePersistenceCoordinator<R>({
+			// Host-routed for stub visibility (see the deps doc comment).
+			canUseNativeBackboneResidentCoordinateState: () =>
+				this.canUseNativeBackboneResidentCoordinateState(),
+			host: () => this,
+			nativeBackbone: () => this._nativeBackbone,
+			nativeRangePlanner: () => this._nativeRangePlanner,
+			nativeSharedLogState: () => this._nativeSharedLogState,
+			entryCoordinatesIndex: () => this.entryCoordinatesIndex,
+			log: () => this.log,
+			remoteBlocks: () => this.remoteBlocks,
+			domain: () => this.domain,
+			indexableDomain: () => this.indexableDomain,
+			coordinateToHash: () => this.coordinateToHash,
+			timeUntilRoleMaturity: () => this.timeUntilRoleMaturity,
+			getEntryGid: (entry) => this.getEntryGid(entry),
+			getEntryNext: (entry) => this.getEntryNext(entry),
+			getEntryHashNumber: (entry) => this.getEntryHashNumber(entry),
+			canPlanNativeHashGid: (entry) => this.canPlanNativeHashGid(entry),
+			hasCustomFindLeaders: () => this.hasCustomFindLeaders(),
+			captureReplicationOwnershipLifecycle: () =>
+				this.captureReplicationOwnershipLifecycle(),
+			throwIfReplicationOwnershipLifecycleInactive: (controller) =>
+				this.throwIfReplicationOwnershipLifecycleInactive(controller),
+			throwIfReplicationOwnershipPoisoned: () =>
+				this.throwIfReplicationOwnershipPoisoned(),
+			isDropStarted: () => this._nativeBackboneDropStarted,
+			getDurableCommitFailure: () => this._nativeDurableCommitFailure,
+			isDurableRecoveryReadyForReopen: () =>
+				this._nativeDurableRecoveryReadyForReopen,
+			setDurableRecoveryReadyForReopen: (value) => {
+				this._nativeDurableRecoveryReadyForReopen = value;
+			},
+		});
+	}
+
 	private ensureNativeDurabilityRuntimeState(): void {
 		// Program clones are borsh-created without running class field initializers.
 		// Keep recovery state from an existing generation, while supplying fresh
@@ -3595,7 +3588,7 @@ export class SharedLog<
 		// Constructing the coordinator supplies the same fresh defaults the
 		// legacy field initializers did (`journal last-flush ??= 0` included):
 		// coordinator instances are always `new`ed, never borsh-cloned.
-		this._coordinates ??= new CoordinatePersistenceCoordinator<R>();
+		this._coordinates ??= this.createCoordinatePersistenceCoordinator();
 		this._nativeStrictDurableDocumentRecoveryDeferred ??= false;
 		this._nativeStrictDurableTransactionsClosing ??= false;
 	}
@@ -5377,90 +5370,25 @@ export class SharedLog<
 		);
 	}
 
+	// Stage 4.5: KEEP-OLD delegators for the early coordinate helpers that
+	// moved to src/coordinate-persistence.ts (the *ForHashValues internal
+	// halves moved without delegators).
 	private deleteCoordinatesForHashes(
 		hashes: Iterable<string>,
 		ownershipLifecycleController?: AbortController,
 	): MaybePromise<void> {
-		if (ownershipLifecycleController) {
-			this.throwIfReplicationOwnershipLifecycleInactive(
-				ownershipLifecycleController,
-			);
-		}
-		const values = normalizedHashValues(hashes);
-		if (values.length === 0) {
-			return;
-		}
-		this.forgetCoordinateStateForHashValues(values);
-		const coordinateIndex = this.entryCoordinatesIndex as PutAndDeleteIndex<
-			EntryReplicated<R>
-		>;
-		if (coordinateIndex.delIdsNoReturn) {
-			return mapMaybePromise(coordinateIndex.delIdsNoReturn(values), () => {
-				if (ownershipLifecycleController) {
-					this.throwIfReplicationOwnershipLifecycleInactive(
-						ownershipLifecycleController,
-					);
-				}
-			});
-		}
-		if (coordinateIndex.delIds) {
-			return mapMaybePromise(coordinateIndex.delIds(values), () => {
-				if (ownershipLifecycleController) {
-					this.throwIfReplicationOwnershipLifecycleInactive(
-						ownershipLifecycleController,
-					);
-				}
-			});
-		}
-		return mapMaybePromise(
-			this.entryCoordinatesIndex.del({
-				query:
-					values.length === 1
-						? { hash: values[0] }
-						: new Or(
-								values.map(
-									(hash) => new StringMatch({ key: "hash", value: hash }),
-								),
-							),
-			}),
-			() => {
-				if (ownershipLifecycleController) {
-					this.throwIfReplicationOwnershipLifecycleInactive(
-						ownershipLifecycleController,
-					);
-				}
-			},
+		return this._coordinates.deleteCoordinatesForHashes(
+			hashes,
+			ownershipLifecycleController,
 		);
 	}
 
 	private forgetCoordinateStateForHashes(hashes: Iterable<string>) {
-		const values = normalizedHashValues(hashes);
-		if (values.length === 0) {
-			return;
-		}
-		this.forgetCoordinateStateForHashValues(values);
-	}
-
-	private forgetCoordinateStateForHashValues(values: string[]) {
-		this._nativeSharedLogState?.deleteEntryCoordinatesBatch(values);
-		this._nativeBackbone?.deleteEntryCoordinatesBatch(values);
-		this.forgetResidentCoordinateStateForHashValues(values);
+		return this._coordinates.forgetCoordinateStateForHashes(hashes);
 	}
 
 	private forgetResidentCoordinateStateForHashes(hashes: Iterable<string>) {
-		const values = normalizedHashValues(hashes);
-		if (values.length === 0) {
-			return;
-		}
-		this.forgetResidentCoordinateStateForHashValues(values);
-	}
-
-	private forgetResidentCoordinateStateForHashValues(values: string[]) {
-		if (this._coordinates._residentEntryCoordinatesByHash) {
-			for (const hash of values) {
-				this._coordinates._residentEntryCoordinatesByHash.delete(hash);
-			}
-		}
+		return this._coordinates.forgetResidentCoordinateStateForHashes(hashes);
 	}
 
 	private async ensureCurrentHeadCoordinatesIndexed(
@@ -20250,73 +20178,41 @@ export class SharedLog<
 		});
 	}
 
+	// Stage 4.5: the coordinate-persistence method cluster moved verbatim to
+	// src/coordinate-persistence.ts (CoordinatePersistenceCoordinator). The
+	// thin delegators below KEEP-OLD every externally-called name so all
+	// legacy call sites stay untouched; instance-level sinon spies/stubs of
+	// these names re-point to the coordinator (`_coordinates`), which every
+	// call — delegated or coordinator-internal — passes through. The
+	// internal-only helpers (nativeBackboneReceiveCoordinateRowsToColumns,
+	// emitNativeBackboneRawCommitProfile,
+	// createCoordinatePersistenceEntry{,FromLeaderPlan},
+	// createCoordinateEntryFromNativeFields,
+	// flushNativeBackboneCoordinateJournalOnAppend,
+	// shouldFlushNativeBackboneCoordinateJournalOnAppend,
+	// canUseRuntimeOnlyNativeBackboneCoordinates,
+	// forget{,Resident}CoordinateStateForHashValues) moved without
+	// delegators.
 	async createCoordinates(
 		entry: ShallowOrFullEntry<any> | EntryReplicated<R> | NumberFromType<R>,
 		minReplicas: number,
-	) {
-		if (
-			typeof entry !== "number" &&
-			typeof entry !== "bigint" &&
-			this.canPlanNativeHashGid(entry)
-		) {
-			const nativeCoordinates = (
-				this._nativeBackbone ?? this._nativeRangePlanner
-			)?.getGidCoordinates(entry.meta.gid, minReplicas) as
-				| NumberFromType<R>[]
-				| undefined;
-			if (nativeCoordinates) {
-				return nativeCoordinates;
-			}
-		}
-
-		const cursor =
-			typeof entry === "number" || typeof entry === "bigint"
-				? entry
-				: await this.domain.fromEntry(entry);
-		const nativeGrid = (
-			this._nativeBackbone ?? this._nativeRangePlanner
-		)?.getGrid(cursor, minReplicas) as NumberFromType<R>[] | undefined;
-		return (
-			nativeGrid ?? this.indexableDomain.numbers.getGrid(cursor, minReplicas)
-		);
+	): Promise<NumberFromType<R>[]> {
+		return this._coordinates.createCoordinates(
+			entry,
+			minReplicas,
+		) as Promise<NumberFromType<R>[]>;
 	}
 
-	private async getCoordinates(entry: { hash: string }) {
-		const nativeCoordinates = (
-			this._nativeBackbone ?? this._nativeSharedLogState
-		)?.getEntryCoordinates(entry.hash);
-		if (nativeCoordinates) {
-			return nativeCoordinates as NumberFromType<R>[];
-		}
-		const result = await this.entryCoordinatesIndex
-			.iterate({ query: { hash: entry.hash } })
-			.all();
-		return result[0].value.coordinates;
+	private async getCoordinates(entry: {
+		hash: string;
+	}): Promise<NumberFromType<R>[]> {
+		return this._coordinates.getCoordinates(entry) as Promise<
+			NumberFromType<R>[]
+		>;
 	}
 
 	private getNativeLogEntryMetadataBatch(hashes: Iterable<string>) {
-		const normalized = [...hashes];
-		if (normalized.length === 0) {
-			return [];
-		}
-		const backboneMetadata =
-			this._nativeBackbone?.graph.entryMetadataHintsBatch(normalized) ??
-			this._nativeBackbone?.graph.entryMetadataBatch(normalized);
-		if (backboneMetadata?.every((entry) => entry != null)) {
-			return backboneMetadata;
-		}
-		const indexMetadata =
-			this.log.entryIndex.getNativeEntryMetadataHintsBatch(normalized) ??
-			this.log.entryIndex.getNativeEntryMetadataBatch(normalized);
-		if (!backboneMetadata) {
-			return indexMetadata;
-		}
-		if (!indexMetadata) {
-			return backboneMetadata;
-		}
-		return backboneMetadata.map(
-			(entry, index) => entry ?? indexMetadata[index],
-		);
+		return this._coordinates.getNativeLogEntryMetadataBatch(hashes);
 	}
 
 	private createReusableReceiveCoordinatePlans(
@@ -20330,250 +20226,38 @@ export class SharedLog<
 			allowRoleAgeZeroPlans?: boolean;
 		},
 	): Map<string, ReusableReceiveCoordinatePlan<R>> {
-		const reusablePlans = new Map<string, ReusableReceiveCoordinatePlan<R>>();
-		if (this.timeUntilRoleMaturity > 0 && !options?.allowRoleAgeZeroPlans) {
-			return reusablePlans;
-		}
-
-		for (const group of receiveGroups) {
-			const plan = group.leaderPlan;
-			if (!plan) {
-				continue;
-			}
-			const replicas =
-				options?.decodedReplicaCounts?.get(group.latestEntry.hash) ??
-				decodeReplicas(group.latestEntry).getValue(this);
-			if (replicas !== group.maxMaxReplicas) {
-				continue;
-			}
-			const prepared = this.createCoordinatePersistenceEntryFromLeaderPlan({
-				entry: group.latestEntry,
-				plan,
-				replicas,
-			});
-			if (!prepared) {
-				continue;
-			}
-			reusablePlans.set(group.latestEntry.hash, {
-				plan,
-				replicas,
-				prepared,
-			});
-		}
-		return reusablePlans;
+		return this._coordinates.createReusableReceiveCoordinatePlans(
+			receiveGroups,
+			options,
+		);
 	}
 
 	private createBackboneOnlyReceiveCoordinateBatch(
 		items: CoordinatePersistBatchItem<R>[],
 	): NativeBackboneReceiveCoordinateBatch<R> | undefined {
-		if (
-			!this._nativeBackbone ||
-			items.length === 0 ||
-			!this.canUseBackboneOnlyCoordinatePersistence()
-		) {
-			return undefined;
-		}
-
-		const rows = items
-			.filter((item) => item.prepared)
-			.map((item) => {
-				const prepared = item.prepared!;
-				const deleteHashes = this.getEntryNext(item.entry);
-				return {
-					item,
-					prepared,
-					fields: prepared.fields,
-					deleteHashes,
-				};
-			});
-		if (rows.length === 0) {
-			return undefined;
-		}
-
-		return {
-			rows,
-			rollbackCoordinateEntries: this.snapshotResidentCoordinateEntries(
-				rows.flatMap((row) => [row.item.entry.hash, ...row.deleteHashes]),
-			),
-		};
-	}
-
-	private nativeBackboneReceiveCoordinateRowsToColumns(
-		rows: NativeBackboneReceiveCoordinateRow<R>[],
-	): NativeBackboneCoordinateCommitColumns {
-		const hashes = new Array<string>(rows.length);
-		const gids = new Array<string>(rows.length);
-		const hashNumberValues = new BigUint64Array(rows.length);
-		const coordinateCounts = new Uint32Array(rows.length);
-		const coordinateValues = new BigUint64Array(
-			rows.reduce((sum, row) => sum + row.fields.coordinates.length, 0),
-		);
-		const nextHashBatches = new Array<string[]>(rows.length);
-		const assignedToRangeBoundaries = new Uint8Array(rows.length);
-		const requestedReplicaValues = new Uint32Array(rows.length);
-		let coordinateOffset = 0;
-		for (let i = 0; i < rows.length; i++) {
-			const { item, prepared, fields, deleteHashes } = rows[i]!;
-			hashes[i] = item.entry.hash;
-			gids[i] = fields.gid;
-			hashNumberValues[i] =
-				typeof fields.hashNumber === "bigint"
-					? fields.hashNumber
-					: BigInt(fields.hashNumberString ?? fields.hashNumber);
-			coordinateCounts[i] = fields.coordinates.length;
-			for (const coordinate of fields.coordinates) {
-				coordinateValues[coordinateOffset++] =
-					typeof coordinate === "bigint" ? coordinate : BigInt(coordinate);
-			}
-			nextHashBatches[i] = deleteHashes;
-			assignedToRangeBoundaries[i] =
-				prepared.assignedToRangeBoundary === true ? 1 : 0;
-			requestedReplicaValues[i] = item.replicas;
-		}
-		return {
-			hashes,
-			gids,
-			hashNumberValues,
-			coordinateCounts,
-			coordinateValues,
-			nextHashBatches,
-			assignedToRangeBoundaries,
-			requestedReplicaValues,
-		};
+		return this._coordinates.createBackboneOnlyReceiveCoordinateBatch(items);
 	}
 
 	private async finishBackboneOnlyReceiveCoordinateBatch(
 		batch: NativeBackboneReceiveCoordinateBatch<R>,
 		profile?: SyncProfileFn,
 	): Promise<Set<string>> {
-		const mirrorStartedAt = syncProfileStart(profile);
-		const persistedHashes = new Set<string>();
-		const coordinateToHashRows: [NumberFromType<R>, string][] = [];
-		let deleteCount = 0;
-		for (const { item, prepared, fields, deleteHashes } of batch.rows) {
-			persistedHashes.add(item.entry.hash);
-			this._coordinates._residentEntryCoordinatesByHash?.set(
-				item.entry.hash,
-				prepared.coordinateEntry ?? fields,
-			);
-			for (const deletedHash of deleteHashes) {
-				this._coordinates._residentEntryCoordinatesByHash?.delete(deletedHash);
-				deleteCount++;
-			}
-			for (const coordinate of item.coordinates) {
-				coordinateToHashRows.push([coordinate, item.entry.hash]);
-			}
-		}
-		this.coordinateToHash.addMany(coordinateToHashRows);
-		emitSyncProfileDuration(profile, mirrorStartedAt, {
-			name: "sharedLog.receive.coordinateResidentMirror",
-			component: "shared-log",
-			entries: batch.rows.length,
-			count: coordinateToHashRows.length,
-			messages: 1,
-			details: { deletes: deleteCount },
-		});
-
-		const flushStartedAt = syncProfileStart(profile);
-		const flushed = this.flushNativeBackboneCoordinateJournalOnAppend();
-		if (isPromiseLike(flushed)) {
-			await flushed;
-		}
-		emitSyncProfileDuration(profile, flushStartedAt, {
-			name: "sharedLog.receive.coordinateJournalFlush",
-			component: "shared-log",
-			entries: batch.rows.length,
-			messages: 1,
-		});
-		return persistedHashes;
+		return this._coordinates.finishBackboneOnlyReceiveCoordinateBatch(
+			batch,
+			profile,
+		);
 	}
 
 	private rollbackBackboneOnlyReceiveCoordinateBatch(
 		batch: NativeBackboneReceiveCoordinateBatch<R>,
 	): void {
-		for (const { item } of batch.rows) {
-			this.rollbackNativeBackboneCoordinateAppend(
-				item.entry.hash,
-				batch.rollbackCoordinateEntries,
-			);
-		}
+		return this._coordinates.rollbackBackboneOnlyReceiveCoordinateBatch(batch);
 	}
 
 	private async persistBackboneOnlyReceiveCoordinateBatch(
 		items: CoordinatePersistBatchItem<R>[],
 	): Promise<Set<string> | undefined> {
-		const backbone = this._nativeBackbone;
-		const batch = this.createBackboneOnlyReceiveCoordinateBatch(items);
-		if (!backbone || !batch) {
-			return undefined;
-		}
-		try {
-			backbone.commitEntryCoordinatesColumnsBatch(
-				this.nativeBackboneReceiveCoordinateRowsToColumns(batch.rows),
-			);
-			return await this.finishBackboneOnlyReceiveCoordinateBatch(batch);
-		} catch (error) {
-			this.rollbackBackboneOnlyReceiveCoordinateBatch(batch);
-			throw error;
-		}
-	}
-
-	private emitNativeBackboneRawCommitProfile(
-		profile: SyncProfileFn | undefined,
-		nativeProfile: NativeBackboneAppendProfile | undefined,
-		entries: number,
-		verifyCount: number,
-	): void {
-		if (!profile || !nativeProfile) {
-			return;
-		}
-		const events: Array<[name: string, durationMs: number, count?: number]> = [
-			[
-				"sharedLog.receive.nativeRawCommit.pendingCheck",
-				nativeProfile.nativeBackboneRawReceivePendingCheckMs,
-			],
-			[
-				"sharedLog.receive.nativeRawCommit.verify",
-				nativeProfile.nativeBackboneRawReceiveVerifyMs,
-				verifyCount,
-			],
-			[
-				"sharedLog.receive.nativeRawCommit.verifyStatus",
-				nativeProfile.nativeBackboneRawReceiveVerifyStatusMs,
-			],
-			[
-				"sharedLog.receive.nativeRawCommit.joinPlan",
-				nativeProfile.nativeBackboneRawReceiveJoinPlanMs,
-			],
-			[
-				"sharedLog.receive.nativeRawCommit.removePending",
-				nativeProfile.nativeBackboneRawReceiveRemoveMs,
-			],
-			[
-				"sharedLog.receive.nativeRawCommit.blockPut",
-				nativeProfile.nativeBackboneRawReceiveBlockPutMs,
-			],
-			[
-				"sharedLog.receive.nativeRawCommit.graphPut",
-				nativeProfile.nativeBackboneRawReceiveGraphPutMs,
-			],
-			[
-				"sharedLog.receive.nativeRawCommit.coordinateCommit",
-				nativeProfile.nativeBackboneRawReceiveCoordinateCommitMs,
-			],
-		];
-		for (const [name, durationMs, count] of events) {
-			if (durationMs > 0) {
-				emitSyncProfileEvent(profile, {
-					name,
-					component: "shared-log",
-					durationMs,
-					entries,
-					count,
-					messages: 1,
-				});
-			}
-		}
+		return this._coordinates.persistBackboneOnlyReceiveCoordinateBatch(items);
 	}
 
 	private createNativeBackbonePreparedJoinCommit(
@@ -20595,245 +20279,14 @@ export class SharedLog<
 				validatePlan?: boolean;
 		  }) => boolean)
 		| undefined {
-		const backbone = this._nativeBackbone;
-		if (
-			!backbone ||
-			this.remoteBlocks?.localStore !== backbone.blocks ||
-			(verifyHashes &&
-				verifyHashes.length > 0 &&
-				!backbone.graph.commitVerifiedPreparedRawReceiveJoinBatch)
-		) {
-			return undefined;
-		}
-		return ({
-			entries,
-			hashes,
-			headFlags,
-			headFlagsBytes,
-			trustedMissing,
-			validatePlan,
-		}) => {
-			this.throwIfReplicationOwnershipPoisoned();
-			if (!trustedMissing || entries.length === 0) {
-				return false;
-			}
-			const coordinateColumns =
-				coordinateBatch && coordinateBatch.rows.length > 0
-					? this.nativeBackboneReceiveCoordinateRowsToColumns(
-							coordinateBatch.rows,
-						)
-					: undefined;
-			if (validatePlan) {
-				const verifiedCommitStartedAt = syncProfileStart(profile);
-				const profileNativeBackbone =
-					!!profile &&
-					!!backbone.resetAppendProfile &&
-					!!backbone.setAppendProfileEnabled &&
-					!!backbone.appendProfile;
-				if (profileNativeBackbone) {
-					backbone.resetAppendProfile();
-					backbone.setAppendProfileEnabled(true);
-				}
-				let committed: boolean | undefined;
-				try {
-					if (verifyHashes && verifyHashes.length > 0) {
-						if (verifyAllHashes) {
-							committed =
-								backbone.graph.commitVerifiedAllPreparedRawReceiveJoinBatch?.(
-									hashes,
-									headFlagsBytes,
-									coordinateColumns,
-								);
-						}
-						committed ??=
-							backbone.graph.commitVerifiedPreparedRawReceiveJoinBatch?.(
-								hashes,
-								headFlagsBytes,
-								verifyHashes,
-								coordinateColumns,
-							);
-					} else {
-						committed = backbone.graph.commitPreparedRawReceiveJoinBatch?.(
-							hashes,
-							headFlagsBytes,
-							coordinateColumns,
-						);
-					}
-				} finally {
-					if (profileNativeBackbone) {
-						backbone.setAppendProfileEnabled(false);
-						this.emitNativeBackboneRawCommitProfile(
-							profile,
-							backbone.appendProfile(),
-							entries.length,
-							verifyHashes?.length ?? 0,
-						);
-					}
-				}
-				if (verifyHashes && verifyHashes.length > 0 && profile) {
-					emitSyncProfileDuration(profile, verifiedCommitStartedAt, {
-						name: "sharedLog.receive.nativeVerifiedCommit",
-						component: "shared-log",
-						entries: entries.length,
-						count: verifyHashes.length,
-						messages: 1,
-					});
-				}
-				if (committed === true) {
-					onPreparedEntriesCommitted?.(hashes);
-					if (coordinateBatch) {
-						onCoordinatesCommitted?.(coordinateBatch);
-					}
-					return true;
-				}
-				if (committed === false) {
-					backbone.graph.clearPreparedRawReceiveEntries?.(hashes);
-					return false;
-				}
-			}
-			if (
-				backbone.graph.commitPreparedRawReceiveBatch(
-					hashes,
-					headFlagsBytes,
-					coordinateColumns,
-				)
-			) {
-				onPreparedEntriesCommitted?.(hashes);
-				if (coordinateBatch) {
-					onCoordinatesCommitted?.(coordinateBatch);
-				}
-				return true;
-			}
-			const commitEntries = new Array<NativeBackboneLogCommitEntry>(
-				entries.length,
-			);
-			for (let i = 0; i < entries.length; i++) {
-				const entry = entries[i]!;
-				if (
-					!entry.bytes ||
-					!entry.nativeEntry ||
-					entry.meta.type !== EntryType.APPEND
-				) {
-					return false;
-				}
-				commitEntries[i] = {
-					...entry.nativeEntry,
-					head: headFlags[i] ?? true,
-					bytes: entry.bytes,
-				};
-			}
-			if (coordinateBatch && coordinateBatch.rows.length > 0) {
-				backbone.graph.commitBlocksGraphAndCoordinatesBatch(
-					commitEntries,
-					coordinateColumns!,
-				);
-				onCoordinatesCommitted?.(coordinateBatch);
-			} else {
-				backbone.graph.commitBlocksAndGraphBatch(commitEntries);
-			}
-			onPreparedEntriesCommitted?.(hashes);
-			return true;
-		};
-	}
-
-	private createCoordinatePersistenceEntryFromLeaderPlan(properties: {
-		entry: ShallowOrFullEntry<any> | EntryReplicated<R>;
-		plan: EntryLeaderPlan<R>;
-		replicas: number;
-	}): PreparedCoordinatePersistence<R> | false {
-		const assignedToRangeBoundary =
-			properties.plan.assignedToRangeBoundary ??
-			shouldAssignToRangeBoundary(properties.plan.leaders, properties.replicas);
-		const hashNumber = this.getEntryHashNumber(properties.entry);
-		const metaBytes = (properties.entry as EntryWithMetaBytes).getMetaBytes?.();
-		if (metaBytes) {
-			const rawTimestamp =
-				properties.entry instanceof Entry
-					? getPreparedRawExchangeTimestamp(properties.entry)
-					: undefined;
-			const wallTime =
-				rawTimestamp?.wallTime ??
-				properties.entry.meta.clock.timestamp.wallTime;
-			return {
-				assignedToRangeBoundary,
-				fields: {
-					hash: properties.entry.hash,
-					hashNumber,
-					hashNumberString: hashNumber.toString(),
-					gid: this.getEntryGid(properties.entry),
-					coordinates: properties.plan.coordinates,
-					coordinateStrings:
-						properties.plan.coordinateStrings ??
-						properties.plan.coordinates.map((coordinate) =>
-							coordinate.toString(),
-						),
-					wallTime,
-					wallTimeString: wallTime.toString(),
-					assignedToRangeBoundary,
-					metaBytes,
-				},
-			};
-		}
-		return this.createCoordinatePersistenceEntry({
-			coordinates: properties.plan.coordinates,
-			entry: properties.entry,
-			leaders: properties.plan.leaders,
-			replicas: properties.replicas,
-			assignedToRangeBoundary,
-			hashNumber,
-		});
-	}
-
-	private createCoordinatePersistenceEntry(properties: {
-		coordinates: NumberFromType<R>[];
-		entry: ShallowOrFullEntry<any> | EntryReplicated<R>;
-		leaders:
-			| Map<
-					string,
-					{
-						intersecting: boolean;
-					}
-			  >
-			| false;
-		replicas: number;
-		prev?: EntryReplicated<R>;
-		assignedToRangeBoundary?: boolean;
-		hashNumber?: NumberFromType<R>;
-	}): PreparedCoordinatePersistence<R> | false {
-		const assignedToRangeBoundary =
-			properties.assignedToRangeBoundary ??
-			shouldAssignToRangeBoundary(properties.leaders, properties.replicas);
-
-		if (
-			properties.prev &&
-			properties.prev.assignedToRangeBoundary === assignedToRangeBoundary
-		) {
-			return false;
-		}
-
-		const metaBytes = (properties.entry as EntryWithMetaBytes).getMetaBytes?.();
-		const coordinateEntry = new this.indexableDomain.constructorEntry({
-			assignedToRangeBoundary,
-			coordinates: properties.coordinates,
-			meta: properties.entry.meta,
-			metaBytes,
-			hash: properties.entry.hash,
-			hashNumber:
-				properties.hashNumber ?? this.getEntryHashNumber(properties.entry),
-		});
-		return {
-			coordinateEntry,
-			assignedToRangeBoundary,
-			fields: {
-				hash: coordinateEntry.hash,
-				hashNumber: coordinateEntry.hashNumber,
-				gid: coordinateEntry.gid,
-				coordinates: coordinateEntry.coordinates,
-				wallTime: coordinateEntry.wallTime,
-				assignedToRangeBoundary: coordinateEntry.assignedToRangeBoundary,
-				metaBytes: coordinateEntry.getMetaBytes(),
-			},
-		};
+		return this._coordinates.createNativeBackbonePreparedJoinCommit(
+			coordinateBatch,
+			onCoordinatesCommitted,
+			verifyHashes,
+			verifyAllHashes,
+			profile,
+			onPreparedEntriesCommitted,
+		);
 	}
 
 	private createCoordinatePersistenceEntryFromNativePlan(properties: {
@@ -20841,230 +20294,53 @@ export class SharedLog<
 		plan: NativeAppendCoordinatePlan;
 		prev?: EntryReplicated<R>;
 	}): PreparedCoordinatePersistence<R> | false {
-		if (
-			properties.plan.hash !== properties.entry.hash ||
-			properties.plan.gid !== this.getEntryGid(properties.entry)
-		) {
-			return false;
-		}
-
-		const assignedToRangeBoundary = properties.plan.assignedToRangeBoundary;
-		if (
-			properties.prev &&
-			properties.prev.assignedToRangeBoundary === assignedToRangeBoundary
-		) {
-			return false;
-		}
-
-		const coordinates = properties.plan.coordinates as NumberFromType<R>[];
-		const hashNumber = properties.plan.hashNumber as NumberFromType<R>;
-		const metaBytes = (properties.entry as EntryWithMetaBytes).getMetaBytes?.();
-		if (metaBytes) {
-			const rawTimestamp =
-				properties.entry instanceof Entry
-					? getPreparedRawExchangeTimestamp(properties.entry)
-					: undefined;
-			const wallTime =
-				rawTimestamp?.wallTime ??
-				properties.entry.meta.clock.timestamp.wallTime;
-			return {
-				assignedToRangeBoundary,
-				fields: {
-					hash: properties.plan.hash,
-					hashNumber,
-					hashNumberString: properties.plan.hashNumberString,
-					gid: properties.plan.gid,
-					coordinates,
-					coordinateStrings: properties.plan.coordinateStrings,
-					wallTime,
-					wallTimeString: wallTime.toString(),
-					assignedToRangeBoundary,
-					metaBytes,
-				},
-			};
-		}
-		const entryMeta = properties.entry.meta;
-		const coordinateEntry = new this.indexableDomain.constructorEntry({
-			assignedToRangeBoundary,
-			coordinates,
-			meta: entryMeta,
-			hash: properties.plan.hash,
-			hashNumber,
-		});
-		return {
-			coordinateEntry,
-			assignedToRangeBoundary,
-			fields: {
-				hash: properties.plan.hash,
-				hashNumber,
-				hashNumberString: properties.plan.hashNumberString,
-				gid: properties.plan.gid,
-				coordinates,
-				coordinateStrings: properties.plan.coordinateStrings,
-				wallTime: coordinateEntry.wallTime,
-				wallTimeString: coordinateEntry.wallTime.toString(),
-				assignedToRangeBoundary,
-				metaBytes: coordinateEntry.getMetaBytes(),
-			},
-		};
-	}
-
-	private createCoordinateEntryFromNativeFields(
-		fields: SharedLogCoordinateNativeFields<R>,
-	): EntryReplicated<R> {
-		return new this.indexableDomain.constructorEntry({
-			assignedToRangeBoundary: fields.assignedToRangeBoundary,
-			coordinates: fields.coordinates,
-			metaBytes: fields.metaBytes,
-			gid: fields.gid,
-			wallTime: fields.wallTime,
-			hash: fields.hash,
-			hashNumber: fields.hashNumber,
-		});
+		return this._coordinates.createCoordinatePersistenceEntryFromNativePlan(
+			properties,
+		);
 	}
 
 	private materializePreparedCoordinateEntry(
 		prepared: PreparedCoordinatePersistence<R>,
 	): EntryReplicated<R> {
-		return (prepared.coordinateEntry ??=
-			this.createCoordinateEntryFromNativeFields(prepared.fields));
+		return this._coordinates.materializePreparedCoordinateEntry(prepared);
 	}
 
 	private materializeResidentCoordinateEntry(
 		entry: ResidentCoordinateEntry<R>,
 	): EntryReplicated<R> {
-		return isEntryReplicated(entry)
-			? entry
-			: this.createCoordinateEntryFromNativeFields(entry);
+		return this._coordinates.materializeResidentCoordinateEntry(entry);
 	}
 
 	private materializeRepairDispatchEntries(
 		entries: ReadonlyMap<string, RepairDispatchEntry<R>>,
 	): Map<string, EntryReplicated<R>> {
-		const materialized = new Map<string, EntryReplicated<R>>();
-		for (const [hash, entry] of entries) {
-			materialized.set(hash, this.materializeResidentCoordinateEntry(entry));
-		}
-		return materialized;
+		return this._coordinates.materializeRepairDispatchEntries(entries);
 	}
 
 	private snapshotResidentCoordinateEntries(
 		hashes: Iterable<string>,
 	): NativeBackboneCoordinateRollback<R> | undefined {
-		const uniqueHashes = new Set([...hashes].filter(Boolean));
-		if (uniqueHashes.size === 0) {
-			return undefined;
-		}
-		const entries = new Map<string, ResidentCoordinateEntry<R>>();
-		const generations = new Map<string, number>();
-		const mutationGenerations = (this._coordinates._nativeCoordinateMutationGenerations ??=
-			new Map());
-		for (const hash of uniqueHashes) {
-			const generation = (mutationGenerations.get(hash) ?? 0) + 1;
-			mutationGenerations.set(hash, generation);
-			generations.set(hash, generation);
-			const entry = this._coordinates._residentEntryCoordinatesByHash?.get(hash);
-			if (entry) {
-				entries.set(hash, entry);
-			}
-		}
-		return { hashes: uniqueHashes, entries, generations };
+		return this._coordinates.snapshotResidentCoordinateEntries(hashes);
 	}
 
 	private rollbackNativeBackboneCoordinateAppend(
 		appendHash: string,
 		rollback?: NativeBackboneCoordinateRollback<R>,
 	): void {
-		const backbone = this._nativeBackbone;
-		if (!backbone) {
-			return;
-		}
-		const hashes = rollback?.hashes ?? new Set([appendHash]);
-		const mutationGenerations = (this._coordinates._nativeCoordinateMutationGenerations ??=
-			new Map());
-		for (const hash of hashes) {
-			const expectedGeneration = rollback?.generations.get(hash);
-			if (
-				expectedGeneration !== undefined &&
-				mutationGenerations.get(hash) !== expectedGeneration
-			) {
-				continue;
-			}
-			backbone.deleteEntryCoordinates(hash);
-			this._nativeSharedLogState?.deleteEntryCoordinates(hash);
-			this._coordinates._residentEntryCoordinatesByHash?.delete(hash);
-			const entry = rollback?.entries.get(hash);
-			if (!entry) {
-				continue;
-			}
-			const fields = isEntryReplicated(entry)
-				? {
-						hash: entry.hash,
-						gid: entry.gid,
-						coordinates: entry.coordinates,
-						assignedToRangeBoundary: entry.assignedToRangeBoundary,
-						hashNumber: entry.hashNumber,
-					}
-				: entry;
-			const requestedReplicas = isEntryReplicated(entry)
-				? decodeReplicas(entry).getValue(this)
-				: fields.coordinates.length;
-			backbone.putEntryCoordinates(
-				fields.hash,
-				fields.gid,
-				fields.coordinates,
-				fields.assignedToRangeBoundary,
-				requestedReplicas,
-				fields.hashNumber,
-			);
-			this._nativeSharedLogState?.putEntryCoordinates(
-				fields.hash,
-				fields.gid,
-				fields.coordinates,
-				fields.assignedToRangeBoundary,
-				requestedReplicas,
-				fields.hashNumber,
-			);
-			this._coordinates._residentEntryCoordinatesByHash?.set(hash, entry);
-		}
+		return this._coordinates.rollbackNativeBackboneCoordinateAppend(
+			appendHash,
+			rollback,
+		);
 	}
 
 	private async rollbackNativeBackboneCoordinateAppendDurably(
 		appendHash: string,
 		rollback?: NativeBackboneCoordinateRollback<R>,
 	): Promise<void> {
-		this.rollbackNativeBackboneCoordinateAppend(appendHash, rollback);
-		const coordinateIndex = this.entryCoordinatesIndex as PutAndDeleteIndex<
-			EntryReplicated<R>
-		>;
-		const hashes = rollback?.hashes ?? new Set([appendHash]);
-		const mutationGenerations = (this._coordinates._nativeCoordinateMutationGenerations ??=
-			new Map());
-		for (const hash of hashes) {
-			const expectedGeneration = rollback?.generations.get(hash);
-			if (
-				expectedGeneration !== undefined &&
-				mutationGenerations.get(hash) !== expectedGeneration
-			) {
-				continue;
-			}
-			const previous = rollback?.entries.get(hash);
-			if (previous) {
-				await coordinateIndex.put(
-					this.materializeResidentCoordinateEntry(previous),
-				);
-			} else if (coordinateIndex.delIds) {
-				await coordinateIndex.delIds([hash]);
-			} else if (coordinateIndex.delIdsNoReturn) {
-				await coordinateIndex.delIdsNoReturn([hash]);
-			} else {
-				await coordinateIndex.del({ query: { hash } });
-			}
-		}
-		const flushed = this.flushNativeBackboneCoordinateJournal();
-		if (isPromiseLike(flushed)) {
-			await flushed;
-		}
+		return this._coordinates.rollbackNativeBackboneCoordinateAppendDurably(
+			appendHash,
+			rollback,
+		);
 	}
 
 	private persistPreparedCoordinate(
@@ -21078,135 +20354,12 @@ export class SharedLog<
 			commitNativeBackbone?: boolean;
 			deleteHashes?: string[];
 		},
-		ownershipLifecycleController = this.captureReplicationOwnershipLifecycle(),
+		ownershipLifecycleController?: AbortController,
 	): MaybePromise<boolean> {
-		this.throwIfReplicationOwnershipLifecycleInactive(
+		return this._coordinates.persistPreparedCoordinate(
+			properties,
 			ownershipLifecycleController,
 		);
-		const { assignedToRangeBoundary, fields } = properties.prepared;
-		const deleteHashes = combineCoordinateDeleteHashes(
-			properties.nextHashes,
-			properties.deleteHashes,
-		);
-		const coordinateIndex = this.entryCoordinatesIndex as PutAndDeleteIndex<
-			EntryReplicated<R>
-		>;
-		let deleteNextOptions: DeleteOptions | undefined;
-		let putResult: MaybePromise<unknown>;
-		if (coordinateIndex.putSharedLogCoordinateFieldsAndDeleteHashesNoReturn) {
-			putResult =
-				coordinateIndex.putSharedLogCoordinateFieldsAndDeleteHashesNoReturn(
-					fields,
-					deleteHashes,
-				);
-		} else if (coordinateIndex.putSharedLogCoordinateFieldsAndDeleteHashes) {
-			putResult = coordinateIndex.putSharedLogCoordinateFieldsAndDeleteHashes(
-				fields,
-				deleteHashes,
-			);
-		} else if (coordinateIndex.putSharedLogCoordinateFieldsAndDeleteIds) {
-			putResult = coordinateIndex.putSharedLogCoordinateFieldsAndDeleteIds(
-				fields,
-				deleteHashes,
-				toId(fields.hash),
-			);
-		} else if (coordinateIndex.putSharedLogCoordinateAndDeleteIds) {
-			const coordinateEntry = this.materializePreparedCoordinateEntry(
-				properties.prepared,
-			);
-			putResult = coordinateIndex.putSharedLogCoordinateAndDeleteIds(
-				coordinateEntry,
-				fields,
-				deleteHashes,
-				toId(fields.hash),
-			);
-		} else if (deleteHashes.length > 0 && coordinateIndex.putAndDeleteIds) {
-			const coordinateEntry = this.materializePreparedCoordinateEntry(
-				properties.prepared,
-			);
-			putResult = coordinateIndex.putAndDeleteIds(
-				coordinateEntry,
-				deleteHashes,
-			);
-		} else {
-			const coordinateEntry = this.materializePreparedCoordinateEntry(
-				properties.prepared,
-			);
-			deleteNextOptions =
-				deleteHashes.length === 0
-					? undefined
-					: deleteHashes.length === 1
-						? { query: { hash: deleteHashes[0] } }
-						: {
-								query: new Or(
-									deleteHashes.map(
-										(x) => new StringMatch({ key: "hash", value: x }),
-									),
-								),
-							};
-			if (deleteNextOptions && coordinateIndex.putAndDelete) {
-				putResult = coordinateIndex.putAndDelete(
-					coordinateEntry,
-					deleteNextOptions,
-				);
-			} else {
-				putResult = this.entryCoordinatesIndex.put(coordinateEntry);
-			}
-		}
-
-		const finish = (): MaybePromise<boolean> => {
-			this.throwIfReplicationOwnershipLifecycleInactive(
-				ownershipLifecycleController,
-			);
-			const nativeDeleteHashes = combineCoordinateDeleteHashes(
-				properties.nextHashes,
-				properties.deleteHashes,
-			);
-			if (properties.commitNative !== false) {
-				this._nativeSharedLogState?.commitEntryCoordinates(
-					properties.hash,
-					fields.gid,
-					properties.coordinates,
-					nativeDeleteHashes,
-					assignedToRangeBoundary,
-					properties.replicas,
-					fields.hashNumber,
-				);
-			}
-			if (properties.commitNativeBackbone !== false) {
-				this._nativeBackbone?.commitEntryCoordinates(
-					properties.hash,
-					fields.gid,
-					properties.coordinates,
-					nativeDeleteHashes,
-					assignedToRangeBoundary,
-					properties.replicas,
-					fields.hashNumber,
-				);
-			}
-			if (this._coordinates._residentEntryCoordinatesByHash) {
-				this._coordinates._residentEntryCoordinatesByHash.set(
-					properties.hash,
-					properties.prepared.coordinateEntry ?? fields,
-				);
-				for (const nextHash of nativeDeleteHashes) {
-					this._coordinates._residentEntryCoordinatesByHash.delete(nextHash);
-				}
-			}
-
-			for (const coordinate of properties.coordinates) {
-				this.coordinateToHash.add(coordinate, properties.hash);
-			}
-
-			if (deleteNextOptions && !coordinateIndex.putAndDelete) {
-				return mapMaybePromise(
-					this.entryCoordinatesIndex.del(deleteNextOptions),
-					() => true,
-				);
-			}
-			return true;
-		};
-		return mapMaybePromise(putResult, finish);
 	}
 
 	private persistPreparedCoordinateNativeTransaction(
@@ -21220,73 +20373,12 @@ export class SharedLog<
 			commitNative?: boolean;
 			commitNativeBackbone?: boolean;
 		},
-		ownershipLifecycleController = this.captureReplicationOwnershipLifecycle(),
+		ownershipLifecycleController?: AbortController,
 	): MaybePromise<boolean> {
-		this.throwIfReplicationOwnershipLifecycleInactive(
+		return this._coordinates.persistPreparedCoordinateNativeTransaction(
+			properties,
 			ownershipLifecycleController,
 		);
-		const { fields } = properties.prepared;
-		const putNative =
-			properties.coordinateIndex
-				.putSharedLogCoordinateFieldsEncodedAndDeleteHashesNoReturn ??
-			properties.coordinateIndex
-				.putSharedLogCoordinateFieldsAndDeleteHashesNoReturn;
-		if (!putNative) {
-			return false;
-		}
-		const putResult = putNative.call(
-			properties.coordinateIndex,
-			fields,
-			combineCoordinateDeleteHashes(
-				properties.nextHashes,
-				properties.deleteHashes,
-			),
-		);
-		const finish = () => {
-			this.throwIfReplicationOwnershipLifecycleInactive(
-				ownershipLifecycleController,
-			);
-			const nativeDeleteHashes = combineCoordinateDeleteHashes(
-				properties.nextHashes,
-				properties.deleteHashes,
-			);
-			if (properties.commitNative !== false) {
-				this._nativeSharedLogState?.commitEntryCoordinates(
-					properties.hash,
-					fields.gid,
-					properties.coordinates,
-					nativeDeleteHashes,
-					properties.prepared.assignedToRangeBoundary,
-					properties.coordinates.length,
-					fields.hashNumber,
-				);
-			}
-			if (properties.commitNativeBackbone !== false) {
-				this._nativeBackbone?.commitEntryCoordinates(
-					properties.hash,
-					fields.gid,
-					properties.coordinates,
-					nativeDeleteHashes,
-					properties.prepared.assignedToRangeBoundary,
-					properties.coordinates.length,
-					fields.hashNumber,
-				);
-			}
-			if (this._coordinates._residentEntryCoordinatesByHash) {
-				this._coordinates._residentEntryCoordinatesByHash.set(
-					properties.hash,
-					properties.prepared.coordinateEntry ?? fields,
-				);
-				for (const nextHash of nativeDeleteHashes) {
-					this._coordinates._residentEntryCoordinatesByHash.delete(nextHash);
-				}
-			}
-			for (const coordinate of properties.coordinates) {
-				this.coordinateToHash.add(coordinate, properties.hash);
-			}
-			return true;
-		};
-		return mapMaybePromise(putResult, finish);
 	}
 
 	private persistBackboneCoordinateFieldsNativeTransaction(
@@ -21298,198 +20390,28 @@ export class SharedLog<
 			deleteHashes: string[];
 			skipGenericTransientCoordinateIndex?: boolean;
 		},
-		ownershipLifecycleController = this.captureReplicationOwnershipLifecycle(),
+		ownershipLifecycleController?: AbortController,
 	): MaybePromise<boolean> {
-		this.throwIfReplicationOwnershipLifecycleInactive(
+		return this._coordinates.persistBackboneCoordinateFieldsNativeTransaction(
+			properties,
 			ownershipLifecycleController,
 		);
-		const { fields } = properties;
-		const useBackboneOnlyCoordinatePersistence =
-			this.canUseBackboneOnlyCoordinatePersistence();
-		const finish = (): MaybePromise<boolean> => {
-			this.throwIfReplicationOwnershipLifecycleInactive(
-				ownershipLifecycleController,
-			);
-			this._nativeSharedLogState?.commitEntryCoordinates(
-				properties.hash,
-				fields.gid,
-				properties.coordinates,
-				properties.deleteHashes,
-				fields.assignedToRangeBoundary,
-				properties.coordinates.length,
-				fields.hashNumber,
-			);
-			if (this._coordinates._residentEntryCoordinatesByHash) {
-				this._coordinates._residentEntryCoordinatesByHash.set(properties.hash, fields);
-				for (const deletedHash of properties.deleteHashes) {
-					this._coordinates._residentEntryCoordinatesByHash.delete(deletedHash);
-				}
-			}
-			for (const coordinate of properties.coordinates) {
-				this.coordinateToHash.add(coordinate, properties.hash);
-			}
-			if (this._coordinates._nativeBackboneCoordinatePersistence) {
-				const flushed = this.flushNativeBackboneCoordinateJournalOnAppend();
-				if (isPromiseLike(flushed)) {
-					return mapMaybePromise(flushed, () => true);
-				}
-			}
-			return true;
-		};
-		if (
-			(properties.skipGenericTransientCoordinateIndex &&
-				this.canUseRuntimeOnlyNativeBackboneCoordinates(
-					properties.coordinateIndex,
-				)) ||
-			useBackboneOnlyCoordinatePersistence
-		) {
-			return finish();
-		}
-
-		const putNative =
-			properties.coordinateIndex
-				.putSharedLogCoordinateFieldsEncodedAndDeleteHashesNoReturn ??
-			properties.coordinateIndex
-				.putSharedLogCoordinateFieldsAndDeleteHashesNoReturn;
-		if (!putNative) {
-			return false;
-		}
-		const putResult = putNative.call(
-			properties.coordinateIndex,
-			fields,
-			properties.deleteHashes,
-		);
-		return mapMaybePromise(putResult, finish);
 	}
 
 	private flushNativeBackboneCoordinateJournal(): MaybePromise<void> {
-		const backbone = this._nativeBackbone;
-		const persistence = this._coordinates._nativeBackboneCoordinatePersistence;
-		if (!backbone || !persistence || this._nativeBackboneDropStarted) {
-			return undefined;
-		}
-		if (
-			backbone.coordinatePendingJournalLength === 0 &&
-			backbone.documentPendingJournalLength === 0 &&
-			backbone.documentSignerPendingJournalLength === 0
-		) {
-			return undefined;
-		}
-		return mapMaybePromise(persistence.flushJournal(backbone), () => {
-			this._coordinates._nativeBackboneCoordinateJournalLastFlushMs = Date.now();
-			return undefined;
-		});
-	}
-
-	private flushNativeBackboneCoordinateJournalOnAppend(): MaybePromise<void> {
-		const backbone = this._nativeBackbone;
-		const persistence = this._coordinates._nativeBackboneCoordinatePersistence;
-		if (!backbone || !persistence || this._nativeBackboneDropStarted) {
-			return undefined;
-		}
-		if (persistence.flushJournalOnAppend) {
-			const flushed = persistence.flushJournalOnAppend(backbone);
-			if (!isPromiseLike(flushed)) {
-				return undefined;
-			}
-			return mapMaybePromise(flushed, () => {
-				return undefined;
-			});
-		}
-		if (!this.shouldFlushNativeBackboneCoordinateJournalOnAppend()) {
-			return undefined;
-		}
-		return this.flushNativeBackboneCoordinateJournal();
-	}
-
-	private shouldFlushNativeBackboneCoordinateJournalOnAppend(): boolean {
-		const persistence = this._coordinates._nativeBackboneCoordinatePersistence;
-		if (!persistence || persistence.flushOnAppend !== false) {
-			return true;
-		}
-		const backbone = this._nativeBackbone;
-		if (!backbone || backbone.coordinatePendingJournalLength === 0) {
-			return false;
-		}
-		if (
-			persistence.flushMaxPendingBytes != null &&
-			backbone.coordinatePendingJournalByteLength >=
-				persistence.flushMaxPendingBytes
-		) {
-			return true;
-		}
-		return (
-			persistence.flushIntervalMs != null &&
-			Date.now() - this._coordinates._nativeBackboneCoordinateJournalLastFlushMs >=
-				persistence.flushIntervalMs
-		);
+		return this._coordinates.flushNativeBackboneCoordinateJournal();
 	}
 
 	private async closeNativeBackboneCoordinatePersistence(): Promise<void> {
-		const persistence = this._coordinates._nativeBackboneCoordinatePersistence;
-		if (!persistence) {
-			return;
-		}
-		if (this._nativeBackboneDropStarted) {
-			// `drop()` owns the durable namespace lifecycle. Never flush the live wasm
-			// journals or invoke an ordinary custom close after its tombstone/erase has
-			// started: a close implementation that rewrites cached state could resurrect
-			// files after a successful terminal drop.
-			return;
-		}
-		if (
-			this._nativeDurableCommitFailure &&
-			!this._nativeDurableRecoveryReadyForReopen
-		) {
-			// The failed native transaction was never published by the lower log.
-			// Its coordinate/document/signer records are still only in the wasm
-			// pending journals. Closing without flushing discards that generation;
-			// the next backbone hydrates the last acknowledged checkpoint.
-			await persistence.close?.();
-			this._nativeDurableRecoveryReadyForReopen = true;
-			return;
-		}
-		await this.flushNativeBackboneCoordinateJournal();
-		await persistence.close?.();
+		return this._coordinates.closeNativeBackboneCoordinatePersistence();
 	}
 
 	private canUseBackboneOnlyCoordinatePersistence(): boolean {
-		return (
-			!!this._coordinates._nativeBackboneCoordinatePersistence &&
-			this.canUseNativeBackboneResidentCoordinateState()
-		);
+		return this._coordinates.canUseBackboneOnlyCoordinatePersistence();
 	}
 
 	private canUseNativeBackboneResidentCoordinateState(): boolean {
-		return (
-			!!this._nativeBackbone &&
-			!!this._coordinates._residentEntryCoordinatesByHash &&
-			!this.hasCustomFindLeaders()
-		);
-	}
-
-	private canUseRuntimeOnlyNativeBackboneCoordinates(
-		coordinateIndex: PutAndDeleteIndex<EntryReplicated<R>>,
-	): boolean {
-		if (
-			!this.canUseNativeBackboneResidentCoordinateState() ||
-			Object.prototype.hasOwnProperty.call(
-				coordinateIndex,
-				"putSharedLogCoordinateFieldsEncodedAndDeleteHashesNoReturn",
-			) ||
-			Object.prototype.hasOwnProperty.call(
-				coordinateIndex,
-				"putSharedLogCoordinateFieldsAndDeleteHashesNoReturn",
-			)
-		) {
-			return false;
-		}
-		const persisted = (
-			coordinateIndex as PutAndDeleteIndex<EntryReplicated<R>> & {
-				persisted?: () => MaybePromise<boolean>;
-			}
-		).persisted?.();
-		return persisted === false;
+		return this._coordinates.canUseNativeBackboneResidentCoordinateState();
 	}
 
 	private async persistCoordinate(
@@ -21514,230 +20436,32 @@ export class SharedLog<
 			nextHashes?: string[];
 			prepared?: PreparedCoordinatePersistence<R>;
 		},
-		ownershipLifecycleController = this.captureReplicationOwnershipLifecycle(),
-	) {
-		this.throwIfReplicationOwnershipLifecycleInactive(
-			ownershipLifecycleController,
-		);
-		const prepared =
-			properties.prepared ?? this.createCoordinatePersistenceEntry(properties);
-		if (!prepared) {
-			return false;
-		}
-		return this.persistPreparedCoordinate(
-			{
-				prepared,
-				hash: properties.entry.hash,
-				nextHashes: properties.nextHashes ?? properties.entry.meta.next,
-				coordinates: properties.coordinates,
-				replicas: properties.replicas,
-				commitNative: properties.commitNative,
-				commitNativeBackbone: properties.commitNativeBackbone,
-				deleteHashes: properties.deleteHashes,
-			},
+		ownershipLifecycleController?: AbortController,
+	): Promise<boolean> {
+		return this._coordinates.persistCoordinate(
+			properties,
 			ownershipLifecycleController,
 		);
 	}
 
 	private async persistCoordinatesBatch(
 		items: CoordinatePersistBatchItem<R>[],
-		ownershipLifecycleController = this.captureReplicationOwnershipLifecycle(),
+		ownershipLifecycleController?: AbortController,
 	): Promise<boolean[]> {
-		this.throwIfReplicationOwnershipLifecycleInactive(
+		return this._coordinates.persistCoordinatesBatch(
+			items,
 			ownershipLifecycleController,
 		);
-		if (items.length === 0) {
-			return [];
-		}
-
-		const prepared = items.map((item) => ({
-			item,
-			prepared: item.prepared ?? this.createCoordinatePersistenceEntry(item),
-		}));
-		const changed = prepared.filter(
-			(
-				entry,
-			): entry is {
-				item: (typeof items)[number];
-				prepared: PreparedCoordinatePersistence<R>;
-			} => entry.prepared !== false,
-		);
-		if (changed.length === 0) {
-			return items.map(() => false);
-		}
-
-		const coordinateIndex = this.entryCoordinatesIndex as PutAndDeleteIndex<
-			EntryReplicated<R>
-		>;
-		const canUseGenericPutBatch =
-			typeof coordinateIndex.putBatch === "function" &&
-			changed.every(({ item }) => item.entry.meta.next.length === 0);
-
-		if (
-			coordinateIndex.putSharedLogCoordinateFieldsAndDeleteHashesBatchNoReturn
-		) {
-			await coordinateIndex.putSharedLogCoordinateFieldsAndDeleteHashesBatchNoReturn(
-				changed.map(({ item, prepared }) => ({
-					fields: prepared.fields,
-					deleteHashes: item.entry.meta.next,
-				})),
-			);
-		} else if (
-			coordinateIndex.putSharedLogCoordinateFieldsAndDeleteHashesBatch
-		) {
-			await coordinateIndex.putSharedLogCoordinateFieldsAndDeleteHashesBatch(
-				changed.map(({ item, prepared }) => ({
-					fields: prepared.fields,
-					deleteHashes: item.entry.meta.next,
-				})),
-			);
-		} else if (coordinateIndex.putSharedLogCoordinateFieldsAndDeleteIdsBatch) {
-			await coordinateIndex.putSharedLogCoordinateFieldsAndDeleteIdsBatch(
-				changed.map(({ item, prepared }) => ({
-					fields: prepared.fields,
-					deleteIds: item.entry.meta.next,
-					id: toId(prepared.fields.hash),
-				})),
-			);
-		} else if (coordinateIndex.putSharedLogCoordinatesAndDeleteIdsBatch) {
-			await coordinateIndex.putSharedLogCoordinatesAndDeleteIdsBatch(
-				changed.map(({ item, prepared }) => ({
-					value: this.materializePreparedCoordinateEntry(prepared),
-					fields: prepared.fields,
-					deleteIds: item.entry.meta.next,
-					id: toId(prepared.fields.hash),
-				})),
-			);
-		} else if (canUseGenericPutBatch) {
-			await coordinateIndex.putBatch!(
-				changed.map(({ prepared }) =>
-					this.materializePreparedCoordinateEntry(prepared),
-				),
-			);
-		} else {
-			const results: boolean[] = [];
-			for (const item of items) {
-				results.push(
-					await this.persistCoordinate(item, ownershipLifecycleController),
-				);
-				this.throwIfReplicationOwnershipLifecycleInactive(
-					ownershipLifecycleController,
-				);
-			}
-			return results;
-		}
-		this.throwIfReplicationOwnershipLifecycleInactive(
-			ownershipLifecycleController,
-		);
-
-		const nativeCoordinateCommits = changed.filter(
-			({ item }) => item.commitNative !== false,
-		);
-		if (nativeCoordinateCommits.length > 0 && this._nativeSharedLogState) {
-			if (this._nativeSharedLogState.commitEntryCoordinatesBatch) {
-				this._nativeSharedLogState.commitEntryCoordinatesBatch(
-					nativeCoordinateCommits.map(({ item, prepared }) => ({
-						hash: item.entry.hash,
-						gid: prepared.fields.gid,
-						coordinates: item.coordinates,
-						nextHashes: item.entry.meta.next,
-						assignedToRangeBoundary: prepared.assignedToRangeBoundary,
-						requestedReplicas: item.replicas,
-						hashNumber: prepared.fields.hashNumber,
-					})),
-				);
-			} else {
-				for (const { item, prepared } of nativeCoordinateCommits) {
-					this._nativeSharedLogState.commitEntryCoordinates(
-						item.entry.hash,
-						prepared.fields.gid,
-						item.coordinates,
-						item.entry.meta.next,
-						prepared.assignedToRangeBoundary,
-						item.replicas,
-						prepared.fields.hashNumber,
-					);
-				}
-			}
-		}
-
-		const nativeBackboneCoordinateCommits = changed.filter(
-			({ item }) => item.commitNativeBackbone !== false,
-		);
-		if (nativeBackboneCoordinateCommits.length > 0 && this._nativeBackbone) {
-			if (this._nativeBackbone.commitEntryCoordinatesBatch) {
-				this._nativeBackbone.commitEntryCoordinatesBatch(
-					nativeBackboneCoordinateCommits.map(({ item, prepared }) => ({
-						hash: item.entry.hash,
-						gid: prepared.fields.gid,
-						coordinates: item.coordinates,
-						nextHashes: item.entry.meta.next,
-						assignedToRangeBoundary: prepared.assignedToRangeBoundary,
-						requestedReplicas: item.replicas,
-						hashNumber: prepared.fields.hashNumber,
-					})),
-				);
-			} else {
-				for (const { item, prepared } of nativeBackboneCoordinateCommits) {
-					this._nativeBackbone.commitEntryCoordinates(
-						item.entry.hash,
-						prepared.fields.gid,
-						item.coordinates,
-						item.entry.meta.next,
-						prepared.assignedToRangeBoundary,
-						item.replicas,
-						prepared.fields.hashNumber,
-					);
-				}
-			}
-		}
-
-		for (const { item, prepared } of changed) {
-			if (this._coordinates._residentEntryCoordinatesByHash) {
-				this._coordinates._residentEntryCoordinatesByHash.set(
-					item.entry.hash,
-					prepared.coordinateEntry ?? prepared.fields,
-				);
-				for (const nextHash of item.entry.meta.next) {
-					this._coordinates._residentEntryCoordinatesByHash.delete(nextHash);
-				}
-			}
-			for (const coordinate of item.coordinates) {
-				this.coordinateToHash.add(coordinate, item.entry.hash);
-			}
-		}
-
-		const changedHashes = new Set(
-			changed.map(({ prepared }) => prepared.fields.hash),
-		);
-		return items.map((item) => changedHashes.has(item.entry.hash));
 	}
 
 	private async deleteCoordinates(
 		properties: { hash: string },
 		ownershipLifecycleController?: AbortController,
-	) {
-		if (ownershipLifecycleController) {
-			this.throwIfReplicationOwnershipLifecycleInactive(
-				ownershipLifecycleController,
-			);
-		}
-		this._nativeSharedLogState?.deleteEntryCoordinates(properties.hash);
-		this._nativeBackbone?.deleteEntryCoordinates(properties.hash);
-		this._coordinates._residentEntryCoordinatesByHash?.delete(properties.hash);
-		const coordinateIndex = this.entryCoordinatesIndex as PutAndDeleteIndex<
-			EntryReplicated<R>
-		>;
-		if (coordinateIndex.delIds) {
-			await coordinateIndex.delIds([properties.hash]);
-		} else {
-			await this.entryCoordinatesIndex.del({ query: properties });
-		}
-		if (ownershipLifecycleController) {
-			this.throwIfReplicationOwnershipLifecycleInactive(
-				ownershipLifecycleController,
-			);
-		}
+	): Promise<void> {
+		return this._coordinates.deleteCoordinates(
+			properties,
+			ownershipLifecycleController,
+		);
 	}
 
 	async getDefaultMinRoleAge(): Promise<number> {
